refactored firebase definitions
