+++ src/components/modals/StatsModal.tsx
@@ -23,12 +23,12 @@ import {
   VERSE_TEXT_BUTTON,
 } from '../../constants/strings'
 import { GameStats } from '../../constants/types'
+import { auth } from '../../lib/firebaseConfig'
 import { shareStatus } from '../../lib/share'
 import { referenceUrl, solutionGameDate, tomorrow } from '../../lib/words'
 import { Histogram } from '../stats/Histogram'
 import { MigrationIntro } from '../stats/MigrationIntro'
 import { StatBar } from '../stats/StatBar'
-import { auth } from './../../lib/firebase'
 import { BaseModal } from './BaseModal'
 
 type Props = {
