+++ src/components/profile/ConfirmEditProfileModal.tsx
@@ -2,13 +2,13 @@ import { User } from 'firebase/auth'
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import { PropToEditEnum } from './../../constants/types'
-import { useAlert } from './../../context/AlertContext'
 import {
   resetForgottenPassword,
   updateFirestoreEmail,
   updateFirestoreUsername,
-} from './../../lib/firebase'
+} from '../../lib/firebaseAuth'
+import { PropToEditEnum } from './../../constants/types'
+import { useAlert } from './../../context/AlertContext'
 import { BaseModal } from './../modals/BaseModal'
 
 type Props = {
