+++ src/components/profile/ForgotPasswordModal.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
+import { resetForgottenPassword } from '../../lib/firebaseAuth'
 import { buttonEnabledClasses } from './../../constants/classes'
-import { resetForgottenPassword } from './../../lib/firebase'
 import { BaseModal } from './../modals/BaseModal'
 
 type Props = {
