+++ src/components/profile/LogOutModal.tsx
@@ -1,4 +1,4 @@
-import { logout } from './../../lib/firebase'
+import { logout } from '../../lib/firebaseAuth'
 import { BaseModal } from './../modals/BaseModal'
 
 type Props = {
