+++ src/pages/GroupLeaderboard.tsx
@@ -3,6 +3,13 @@ import { useEffect, useState } from 'react'
 import { useAuthState } from 'react-firebase-hooks/auth'
 import { Link, useNavigate, useParams } from 'react-router-dom'
 
+import {
+  acceptJoinPrivateGroup,
+  denyJoinPrivateGroup,
+  getCleanedGroupName,
+  getGroupLeaderboardByGroupNameFromFirestore,
+  getGroupsByUidFromFirestore,
+} from '../lib/firebaseGroups'
 import Loading from './../components/gameState/Loading'
 import { ConfirmJoinGroupModal } from './../components/groups/ConfirmJoinGroupModal'
 import { ConfirmLeaveGroupModal } from './../components/groups/ConfirmLeaveGroupModal'
@@ -13,14 +20,7 @@ import { buttonEnabledClasses } from './../constants/classes'
 import { Group } from './../constants/types'
 import { useAlert } from './../context/AlertContext'
 import favicon from './../img/favicon.png'
-import {
-  acceptJoinPrivateGroup,
-  auth,
-  denyJoinPrivateGroup,
-  getCleanedGroupName,
-  getGroupLeaderboardByGroupNameFromFirestore,
-  getGroupsByUidFromFirestore,
-} from './../lib/firebase'
+import { auth } from './../lib/firebaseConfig'
 
 function GroupLeaderboard() {
   const params = useParams()
